Radar Warning Integrated
A RDR CONTACT warning has been added to the righ lower screen. If the radar detect a Dynamic Core it will come out and if multiple contacts the distance to the closer one will be displayed. This warning is for situation awarness and an hint for the pilot to check the Radar Widget (ALT+8) if needed.