Minor Update
Removed the Widgets when sit, can be show/hide with ALT+9 and ALT+8 for the radar. In Lua Parameter can flag or deflag if to start with them or no.
Updated the Autobrake logic to work also as Parking Brake. The Parking Brake can be turned ON and OFF any time if LOAD or AI mode are active, if SPC Mode is active it will ARM/Engage the Autobrake function, If in ORB or DMG mode the system will ask to switch to another Mode to activate the Autobrake/Parking Brake.